Move both of DoorInteraction patches to "InteractionPatches"
